dev-requirements.txt
# this should be installed only on development environment
pylint
autopep8
pylint-flask-sqlalchemy
coverage